backend/pyproject.toml
[project]
name = "recircula-backend"
version = "0.1.0"
description = "Plataforma de Doação e Troca – Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.2.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.22",
    "pillow>=12.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]
